+++ src/Enum/ParticipationType.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * This file is part of the package demosplan.
+ *
+ * (c) 2010-present DEMOS plan GmbH, for more information see the license file.
+ *
+ * All rights reserved
+ */
+
+namespace DemosEurope\DemosplanAddon\XBeteiligung\Enum;
+
+/**
+ * XBeteiligung participation types.
+ *
+ * Defines the different types of participants in XBeteiligung procedures.
+ */
+enum ParticipationType: string
+{
+    case INSTITUTION = 'institution';
+    case PUBLIC = 'public';
+}